game loop, player movement, other stuff
-added game loop
-player can walk around
-room has random layout
-made a mess, need to clean up when time is available

+++ CppRogueLike/Dungeon.h
@@ -5,10 +5,10 @@ class Dungeon
 {
 	int width_;
 	int height_;
-	size_t number_of_layers_;
+	int number_of_layers_;
 	Layer* layers_;
 public:
-	Dungeon(int width, int height, size_t layer_count) : width_{ width }, height_{ height }, number_of_layers_{ layer_count }, layers_{ new Layer[number_of_layers_] }
+	Dungeon(int width, int height, int layer_count) : width_{ width }, height_{ height }, number_of_layers_{ layer_count }, layers_{ new Layer[number_of_layers_] }
 	{
 	}
